Refactor(plugins): Rebase and adapt to new structure
