+++ ansible_collections/arista/avd/plugins/action/yaml_templates_to_facts.py
@@ -17,7 +17,7 @@ from ansible_collections.arista.avd.plugins.plugin_utils.avdfacts import AvdFact
 from ansible_collections.arista.avd.plugins.plugin_utils.errors import AristaAvdError
 from ansible_collections.arista.avd.plugins.plugin_utils.merge import merge
 from ansible_collections.arista.avd.plugins.plugin_utils.strip_empties import strip_null_from_data
-from ansible_collections.arista.avd.plugins.plugin_utils.utils import compile_searchpath
+from ansible_collections.arista.avd.plugins.plugin_utils.utils import compile_searchpath, load_python_class
 from ansible_collections.arista.avd.plugins.plugin_utils.utils import template as templater
 
 DEFAULT_PYTHON_CLASS_NAME = "AvdStructuredConfig"
