+++ ansible_collections/arista/avd/plugins/plugin_utils/utils/__init__.py
@@ -3,6 +3,7 @@ from .default import default
 from .get import get
 from .get_all import get_all
 from .get_item import get_item
+from .load_python_class import load_python_class
 from .template import template
 from .template_var import template_var
 from .unique import unique
